core: update ReproduceRequest command to reflect recent workflows
