+++ core/src/main/java/fr/sncf/osrd/cli/BenchSTDCM.kt
@@ -76,10 +76,26 @@ class BenchSTDCM : CliCommand {
     )
     private var railjson: String? = null
     @Parameter(
-        names = ["--timetable-dir"],
-        description = "Path to the timetable directory, must contain timetable_id.json",
+        names = ["--json-timetable"],
+        description =
+            "Path to a JSON file containing the requirements, generated by " +
+                "`download-stdcm-requirements.py`. Using the cbor file is recommended if available.",
     )
-    private var timetableDirectory: String? = null
+    private var jsonTimetable: String? = null
+    @Parameter(
+        names = ["--cbor-timetable"],
+        description =
+            "Path to a cbor file containing the requirements, can be cached by `TimetableCacheManager` or fetched from the s3",
+    )
+    private var cborTimetable: String? = null
+    @Parameter(
+        names = ["--timetable-cache-directory"],
+        description =
+            "Path to a directory to use as timetable cache. " +
+                "Same as the LOCAL_TIMETABLE_CACHE env variable in WorkerCommand. " +
+                "Not necessary when a cbor file is directly provided.",
+    )
+    private var timetableCacheDirectory: String? = null
     private val logger: Logger = LoggerFactory.getLogger("STDCM bench")
 
     override fun run(): Int {
@@ -94,12 +110,13 @@ class BenchSTDCM : CliCommand {
                     FileInfraProvider(infra)
                 } else InfraManager(editoastUrl, editoastAuthorization, httpClient)
             val timetableProvider =
-                if (timetableDirectory != null) JsonTimetableProvider(timetableDirectory!!)
+                if (cborTimetable != null) CBORTimetableReader(cborTimetable!!)
+                else if (jsonTimetable != null) JSONTimetableReader(jsonTimetable!!)
                 else TimetableDownloader(editoastUrl, editoastAuthorization, httpClient)
             val cacheManager =
                 TimetableCacheManager(
                     timetableProvider,
-                    timetableDirectory,
+                    timetableCacheDirectory,
                     osrdGitDescribe = "development",
                 )
