+++ core/src/main/java/fr/sncf/osrd/cli/ReproduceRequest.kt
@@ -2,8 +2,15 @@ package fr.sncf.osrd.cli
 
 import com.beust.jcommander.Parameter
 import com.beust.jcommander.Parameters
+import com.google.common.collect.Range
+import com.google.common.collect.RangeSet
+import com.google.common.collect.TreeRangeSet
 import com.squareup.moshi.JsonAdapter
+import com.squareup.moshi.JsonReader
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import fr.sncf.osrd.api.*
+import fr.sncf.osrd.api.conflicts.TrainRequirementsById
 import fr.sncf.osrd.api.pathfinding.PathfindingBlocksEndpoint
 import fr.sncf.osrd.api.pathfinding.pathfindingRequestAdapter
 import fr.sncf.osrd.api.standalone_sim.SimulationEndpoint
@@ -11,12 +18,19 @@ import fr.sncf.osrd.api.standalone_sim.SimulationRequest
 import fr.sncf.osrd.api.stdcm.STDCMEndpoint
 import fr.sncf.osrd.api.stdcm.stdcmRequestAdapter
 import fr.sncf.osrd.cli.ValidateInfra.parseRailJSONFromFile
+import fr.sncf.osrd.conflicts.SpacingRequirement
+import fr.sncf.osrd.sim_infra.api.RawInfra
+import fr.sncf.osrd.sim_infra.api.ZoneId
 import fr.sncf.osrd.utils.jacoco.ExcludeFromGeneratedCodeCoverage
+import fr.sncf.osrd.utils.json.UnitAdapterFactory
 import java.io.IOException
 import java.nio.file.Path
 import java.util.concurrent.TimeUnit
+import kotlin.io.path.Path
+import kotlin.io.path.readBytes
 import kotlin.time.measureTime
 import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.cbor.Cbor
 import okhttp3.OkHttpClient
 import okio.buffer
 import okio.source
@@ -45,7 +59,8 @@ class ReproduceRequest : CliCommand {
 
     @Parameter(
         names = ["--editoast-url"],
-        description = "The base URL of editoast (used to query infrastructures)",
+        description =
+            "Editoast URL, used to query infrastructures and timetables if not given as files",
     )
     private var editoastUrl = "http://localhost:8090/"
 
@@ -60,10 +75,26 @@ class ReproduceRequest : CliCommand {
     )
     private var railjson: String? = null
     @Parameter(
-        names = ["--timetable-dir"],
-        description = "Path to the timetable directory, must contain timetable_id.json",
+        names = ["--json-timetable"],
+        description =
+            "Path to a JSON file containing the requirements, generated by " +
+                "`download-stdcm-requirements.py`. Using the cbor file is recommended if available.",
     )
-    private var timetableDirectory: String? = null
+    private var jsonTimetable: String? = null
+    @Parameter(
+        names = ["--cbor-timetable"],
+        description =
+            "Path to a cbor file containing the requirements, can be cached by `TimetableCacheManager` or fetched from the s3",
+    )
+    private var cborTimetable: String? = null
+    @Parameter(
+        names = ["--timetable-cache-directory"],
+        description =
+            "Path to a directory to use as timetable cache. " +
+                "Same as the LOCAL_TIMETABLE_CACHE env variable in WorkerCommand. " +
+                "Not necessary when a cbor file is directly provided.",
+    )
+    private var timetableCacheDirectory: String? = null
     private val logger: Logger = LoggerFactory.getLogger("ReproduceRequest")
 
     @OptIn(ExperimentalSerializationApi::class)
@@ -80,12 +111,13 @@ class ReproduceRequest : CliCommand {
                     FileInfraProvider(infra)
                 } else InfraManager(editoastUrl, editoastAuthorization, httpClient)
             val timetableProvider =
-                if (timetableDirectory != null) JsonTimetableProvider(timetableDirectory!!)
+                if (cborTimetable != null) CBORTimetableReader(cborTimetable!!)
+                else if (jsonTimetable != null) JSONTimetableReader(jsonTimetable!!)
                 else TimetableDownloader(editoastUrl, editoastAuthorization, httpClient)
             val cacheManager =
                 TimetableCacheManager(
                     timetableProvider,
-                    timetableDirectory,
+                    timetableCacheDirectory,
                     osrdGitDescribe = "development",
                 )
 
@@ -129,5 +161,85 @@ class ReproduceRequest : CliCommand {
 data class FileInfraProvider(val infra: FullInfra) : InfraProvider {
     override fun getInfra(infraId: String?, expectedVersion: Int?): FullInfra {
         return infra
+    }
+}
+
+/**
+ * Reads a JSON file containing just the timetable requirement, the (concatenated) results of the
+ * `api/timetable/{timetable_id}/requirements/` endpoint. Can be extracted from a timetable using
+ * the `download-stdcm-requirements.py` script.
+ *
+ * Using the serialized .cbor files is more efficient when they're available, but this can be used
+ * even when no .cbor file was saved during the initial processing.
+ */
+class JSONTimetableReader(val jsonTimetablePath: String) : TimetableProvider {
+    private val logger: Logger = LoggerFactory.getLogger("JSONTimetableReader")
+
+    override fun getTimetableData(
+        infraId: String,
+        infra: RawInfra,
+        timetableId: TimetableId,
+    ): STDCMTimetableData {
+        val adapter: JsonAdapter<TrainRequirementsById> =
+            Moshi.Builder()
+                .addLast(UnitAdapterFactory())
+                .addLast(KotlinJsonAdapterFactory())
+                .build()
+                .adapter(TrainRequirementsById::class.java)
+        logger.info("Fetching timetable requirements at json file $jsonTimetablePath")
+
+        val source = Path(jsonTimetablePath).source().buffer()
+        val jsonReader = JsonReader.of(source)
+        jsonReader.beginArray()
+        val zoneUses = mutableMapOf<ZoneId, RangeSet<Double>>()
+        val detailedRequirements =
+            mutableMapOf<ZoneId, MutableList<STDCMTimetableData.DetailedRequirement>>()
+        while (jsonReader.hasNext()) {
+            val train = adapter.fromJson(jsonReader)!!
+            for (rjsSpacingReq in train.spacingRequirements) {
+                val spacingReq =
+                    SpacingRequirement.fromRJSWithAddedTime(
+                        rjsSpacingReq,
+                        infra,
+                        train.startTime.durationSinceEpoch(),
+                    )
+                val set = zoneUses.computeIfAbsent(spacingReq.zone) { TreeRangeSet.create() }
+                set.add(Range.closedOpen(spacingReq.beginTime, spacingReq.endTime))
+                detailedRequirements
+                    .computeIfAbsent(spacingReq.zone) { mutableListOf() }
+                    .add(
+                        STDCMTimetableData.DetailedRequirement(
+                            spacingReq.beginTime,
+                            spacingReq.endTime,
+                            train.trainName,
+                        )
+                    )
+            }
+        }
+
+        return STDCMTimetableData(zoneUses, detailedRequirements)
+    }
+}
+
+/**
+ * Reads a CBOR file with the serialized timetable. Can be found in either the local timetable cache
+ * directory, or in the s3.
+ */
+@OptIn(ExperimentalSerializationApi::class)
+class CBORTimetableReader(val cborTimetablePath: String) : TimetableProvider {
+    private val logger: Logger = LoggerFactory.getLogger("JSONTimetableReader")
+
+    override fun getTimetableData(
+        infraId: String,
+        infra: RawInfra,
+        timetableId: TimetableId,
+    ): STDCMTimetableData {
+        val cbor = Cbor {}
+        val serializer = STDCMTimetableData.SerializableMap.serializer()
+
+        logger.debug("read timetable data from serialized file at $cborTimetablePath")
+        val bytes = Path(cborTimetablePath).readBytes()
+        val serializableMap = cbor.decodeFromByteArray(serializer, bytes)
+        return serializableMap.toSTDCMRequirements()
     }
 }
