+++ core/src/test/java/fr/sncf/osrd/api/WorkerLoadingTest.kt
@@ -1,6 +1,7 @@
 package fr.sncf.osrd.api
 
 import fr.sncf.osrd.api.WorkerLoadEndpoint.WorkerLoadRequest
+import fr.sncf.osrd.cli.JSONTimetableReader
 import fr.sncf.osrd.cli.RqFake
 import java.io.IOException
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -20,8 +21,7 @@ class WorkerLoadingTest : ApiTest() {
         val request =
             if (isRequestNull) null else WorkerLoadRequest("tiny_infra/infra.json", 1, null)
         val requestBody = WorkerLoadEndpoint.adapterRequest.toJson(request)
-        val dummyCacheManager =
-            TimetableCacheManager(JsonTimetableProvider(""), osrdGitDescribe = "")
+        val dummyCacheManager = TimetableCacheManager(JSONTimetableReader(""), osrdGitDescribe = "")
         val response = WorkerLoadEndpoint(infraManager, dummyCacheManager).act(RqFake(requestBody))
         Assertions.assertEquals(expectedStatusCode, response.statusCode())
     }
